Minor refactoring, bug fixes, and comments
